Add initial database JSON file.
This commit adds an initial JSON file for the database. The file
contains posts, comments, and a profile. It serves as a starting point
for the database structure and data. This will allow us to easily
bootstrap and develop the JSON-server with pre-populated data. The
addition of this file simplifies the setup process and saves developers
time when working with the project.

Closes #123.