feat: Improve user experience and add profile pictures to posts
This commit adds profile pictures to the posts displayed on the home
page, enhancing the visual appeal and personalization of the user
experience. It also includes a new "Create Post" feature, which allows
users to upload images and add descriptions for their posts. By
introducing these changes, we aim to create a more engaging and
immersive environment for our users.

Fixes: #1234